+++ keyboards/splitkb/aurora/sweep/keymaps/flinguenheld/rules.mk
@@ -1,11 +1,6 @@
-# Recommended by the doc to improve compilation :
 LTO_ENABLE = yes
 SPLIT_KEYBOARD = yes
-
-
-BOOTLOADER = rp2040
 CONVERT_TO = elite_pi
-
 
 OLED_ENABLE = yes
 OLED_DRIVER = SSD1306
